*) Updated, support flash string printf; *) Add option "--string_literals_in_flash" to C/C++ compiler, to force put the string into flash, to save more SRAM memory;     --> ref: http://www.ourdev.cn/bbs/bbs_content.jsp?bbs_sn=823285&bbs_page_no=2&search_mode=4&search_text=steven&bbs_id=9999

+++ avr-base/inc/typedef.h
@@ -40,8 +40,8 @@ typedef unsigned char           BOOL;       /* Boolean: TRUE/FALSE */
 typedef unsigned char           BYTE;
 typedef unsigned char           UINT8;      /*  8-bit */
 typedef signed   char           SINT8;
-typedef unsigned short          UINT16;     /* 16-bit */
-typedef signed   short          SINT16;
+typedef unsigned int            UINT16;     /* 16-bit */
+typedef signed   int            SINT16;
 typedef unsigned long           UINT32;     /* 32-bit */
 typedef signed   long           SINT32;
 typedef unsigned long long      UINT64;     /* 64-bit */
